Fix conflict between a11y and ui regression tests
