+++ tests/integration/__image_snapshots__/regression-test-js-regression-simple-editor-renders-1-snap.png
